Update AssetThumbnailGenerator now separated into 2 sections Systems section controls thumbnail generation queue Visuals section takes thumbnail pictures Added generate thumbnail for scene assets
