+++ Assets/Scripts/System/AssetThumbnailGeneratorSystem.cs.meta
@@ -1,11 +1,3 @@
 fileFormatVersion: 2
-guid: a7ff8da9234d00744ba88ddf20256d86
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: 974619e3be304493b7870af3ded09254
+timeCreated: 1685456292
