+++ Assets/Scripts/Visuals/AssetThumbnailGenerator.cs.meta
@@ -1,0 +1,11 @@
+﻿fileFormatVersion: 2
+guid: a7ff8da9234d00744ba88ddf20256d86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
